requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
CinetPay==0.0.1
colorlog==6.9.0
cryptography==46.0.2
Django==5.2.7
django-allauth==65.12.0
django-appconf==1.1.0
django-axes==8.0.0
django-ckeditor-5==0.2.18
django-environ==0.12.0
django-filter==25.2
django-humanize==0.1.2
django-imagekit==5.0.0
django-js-asset==3.1.2
django-mptt==0.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
idna==3.10
pilkit==3.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
requests==2.32.5
simplejson==3.20.2
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
